feat(pricing): add usage-aware dry run
