+++ src/core/pricing_service/tests.rs
@@ -246,6 +246,43 @@ fn flat_image_model_info(output_cost_per_image: Option<f64>) -> LiteLLMModelInfo
 }
 
 #[test]
+fn provider_pricing_dry_run_charges_audio_tokens_with_audio_price() {
+    let service = PricingService::new(None);
+    let mut model_info = flat_image_model_info(None);
+    model_info.mode = "audio_transcription".to_string();
+    model_info.extra.insert(
+        "input_cost_per_audio_token".to_string(),
+        serde_json::Value::from(0.002),
+    );
+    service.add_custom_model("audio-priced-model".to_string(), model_info);
+    let mut usage = PricingUsage::new(0, 0);
+    usage.audio_tokens = Some(250);
+
+    let cost = service
+        .dry_run_loaded_usage_cost_for_provider("bedrock", "audio-priced-model", &usage)
+        .unwrap();
+
+    assert_eq!(cost.audio_cost, 0.5);
+    assert_eq!(cost.total_cost, 0.5);
+}
+
+#[test]
+fn provider_pricing_dry_run_fails_closed_for_missing_audio_price() {
+    let service = PricingService::new(None);
+    let mut model_info = flat_image_model_info(None);
+    model_info.mode = "audio_transcription".to_string();
+    service.add_custom_model("missing-audio-price".to_string(), model_info);
+    let mut usage = PricingUsage::new(0, 0);
+    usage.audio_tokens = Some(250);
+
+    let error = service
+        .dry_run_loaded_usage_cost_for_provider("bedrock", "missing-audio-price", &usage)
+        .unwrap_err();
+
+    assert!(error.to_string().contains("input_cost_per_audio_token"));
+}
+
+#[test]
 fn provider_pricing_charges_flat_output_image_cost_without_token_prices() {
     let service = PricingService::new(None);
     service.add_custom_model(
@@ -253,10 +290,11 @@ fn provider_pricing_charges_flat_output_image_cost_without_token_prices() {
         flat_image_model_info(Some(0.06)),
     );
     let mut usage = PricingUsage::new(25, 0);
+    usage.image_tokens = Some(300);
     usage.output_image_count = Some(3);
 
     let cost = service
-        .calculate_loaded_usage_cost_for_provider("bedrock", "flat-image-model", &usage)
+        .dry_run_loaded_usage_cost_for_provider("bedrock", "flat-image-model", &usage)
         .unwrap();
 
     assert_eq!(cost.input_cost, 0.0);
@@ -275,7 +313,7 @@ fn provider_pricing_fails_closed_for_missing_flat_output_image_price() {
     usage.output_image_count = Some(1);
 
     let error = service
-        .calculate_loaded_usage_cost_for_provider("bedrock", "missing-image-price", &usage)
+        .dry_run_loaded_usage_cost_for_provider("bedrock", "missing-image-price", &usage)
         .unwrap_err();
 
     assert!(error.to_string().contains("output_cost_per_image"));
@@ -341,7 +379,7 @@ fn provider_pricing_treats_explicit_zero_image_token_price_as_present() {
 }
 
 #[test]
-fn provider_pricing_does_not_require_flat_price_for_token_priced_image_model() {
+fn provider_pricing_fails_closed_for_token_priced_image_usage_without_image_price() {
     let service = PricingService::new(None);
     let mut model_info = flat_image_model_info(None);
     model_info.input_cost_per_token = Some(0.01);
@@ -351,12 +389,13 @@ fn provider_pricing_does_not_require_flat_price_for_token_priced_image_model() {
     usage.image_tokens = Some(100);
     usage.output_image_count = Some(1);
 
-    let cost = service
-        .calculate_loaded_usage_cost_for_provider("bedrock", "token-priced-image-model", &usage)
-        .unwrap();
+    let error = service
+        .dry_run_loaded_usage_cost_for_provider("bedrock", "token-priced-image-model", &usage)
+        .unwrap_err();
 
-    assert!((cost.input_cost - 0.02).abs() < f64::EPSILON);
-    assert_eq!(cost.image_cost, 0.0);
+    let error = error.to_string();
+    assert!(error.contains("image_cost_per_token"));
+    assert!(error.contains("output_cost_per_image"));
 }
 
 #[test]
